+++ packages/components/src/components/checkbox/checkbox.spec.ts
@@ -96,7 +96,7 @@ describe('componentDidRender', () => {
     expect(setValiditySpy).toHaveBeenCalledTimes(1);
     expect(setValiditySpy).toHaveBeenCalledWith(
       component['checkboxInputElement'].validity,
-      component['checkboxInputElement'].validationMessage,
+      component['checkboxInputElement'].validationMessage || ' ',
       component['checkboxInputElement']
     );
   });