+++ packages/components/src/components/input-email/input-email.tsx
@@ -94,7 +94,7 @@ export class InputEmail {
   @Prop() public placeholder?: string = '';
 
   /** A boolean value that, if present, makes the input field unusable and unclickable. The value will not be submitted with the form. */
-  @Prop() public disabled?: boolean = false;
+  @Prop({ mutable: true }) public disabled?: boolean = false;
 
   /** A boolean value that, if present, indicates that the input field must be filled out before the form can be submitted. */
   @Prop() public required?: boolean = false;
@@ -163,6 +163,7 @@ export class InputEmail {
   }
 
   public formDisabledCallback(disabled: boolean): void {
+    // Called when a parent fieldset is disabled or enabled
     this.disabled = disabled;
   }
 
@@ -179,7 +180,13 @@ export class InputEmail {
   }
 
   public componentDidRender(): void {
-    this.internals?.setValidity(this.inputElement.validity, this.inputElement.validationMessage, this.inputElement);
+    if (!this.disabled && !this.readOnly) {
+      this.internals?.setValidity(
+        this.inputElement.validity,
+        this.inputElement.validationMessage || ' ',
+        this.inputElement
+      );
+    }
   }
 
   public render(): JSX.Element {